+++ organizations_policy.tf
@@ -1,10 +1,10 @@
 locals {
   enabled_root_policies = {
     allowed_regions = {
-      enable = var.aws_root_scp_exceptions.allowed_regions != null ? true : false
-      policy = var.aws_root_scp_exceptions.allowed_regions != null != null ? templatefile("${path.module}/files/organizations/allowed_regions.json.tpl", {
-        allowed    = var.aws_root_scp_exceptions.allowed_regions != null ? var.aws_root_scp_exceptions.allowed_regions : []
-        exceptions = var.aws_root_scp_exceptions.principal_exceptions != null ? var.aws_root_scp_exceptions.principal_exceptions : []
+      enable = var.aws_service_control_policies.allowed_regions != null ? true : false
+      policy = var.aws_service_control_policies.allowed_regions != null != null ? templatefile("${path.module}/files/organizations/allowed_regions.json.tpl", {
+        allowed    = var.aws_service_control_policies.allowed_regions != null ? var.aws_service_control_policies.allowed_regions : []
+        exceptions = var.aws_service_control_policies.principal_exceptions != null ? var.aws_service_control_policies.principal_exceptions : []
       }) : null
     }
     cloudtrail_log_stream = {
@@ -12,20 +12,20 @@ locals {
       policy = file("${path.module}/files/organizations/cloudtrail_log_stream.json")
     }
     deny_disabling_security_hub = {
-      enable = var.aws_deny_disabling_security_hub != false ? true : false
-      policy = var.aws_deny_disabling_security_hub != false ? templatefile("${path.module}/files/organizations/deny_disabling_security_hub.json.tpl", {
-        exceptions = var.aws_root_scp_exceptions.principal_exceptions != null ? var.aws_root_scp_exceptions.principal_exceptions : []
+      enable = var.aws_service_control_policies.aws_deny_disabling_security_hub != false ? true : false
+      policy = var.aws_service_control_policies.aws_deny_disabling_security_hub != false ? templatefile("${path.module}/files/organizations/deny_disabling_security_hub.json.tpl", {
+        exceptions = var.aws_service_control_policies.principal_exceptions != null ? var.aws_service_control_policies.principal_exceptions : []
       }) : null
     }
     deny_leaving_org = {
-      enable = var.aws_deny_leaving_org != false ? true : false
-      policy = var.aws_deny_leaving_org != false ? templatefile("${path.module}/files/organizations/deny_leaving_org.json.tpl", {
-        exceptions = var.aws_root_scp_exceptions.principal_exceptions != null ? var.aws_root_scp_exceptions.principal_exceptions : []
+      enable = var.aws_service_control_policies.aws_deny_leaving_org != false ? true : false
+      policy = var.aws_service_control_policies.aws_deny_leaving_org != false ? templatefile("${path.module}/files/organizations/deny_leaving_org.json.tpl", {
+        exceptions = var.aws_service_control_policies.principal_exceptions != null ? var.aws_service_control_policies.principal_exceptions : []
       }) : null
     }
     // https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ExamplePolicies_EC2.html#iam-example-instance-metadata-requireIMDSv2
     require_use_of_imdsv2 = {
-      enable = var.aws_require_imdsv2
+      enable = var.aws_service_control_policies.aws_require_imdsv2
       policy = file("${path.module}/files/organizations/require_use_of_imdsv2.json")
     }
   }
@@ -49,6 +49,10 @@ resource "aws_organizations_policy" "lz_root_policies" {
   tags        = var.tags
 }
 
+output "root_policy" {
+  value = aws_organizations_policy.lz_root_policies
+}
+
 resource "aws_organizations_policy_attachment" "lz_root_policies" {
   policy_id = aws_organizations_policy.lz_root_policies.id
   target_id = data.aws_organizations_organization.default.roots.0.id
@@ -56,7 +60,7 @@ resource "aws_organizations_policy_attachment" "lz_root_policies" {
 
 // https://summitroute.com/blog/2020/03/25/aws_scp_best_practices/#deny-ability-to-leave-organization
 resource "aws_organizations_policy" "deny_root_user" {
-  count   = length(var.aws_deny_root_user_ous) > 0 ? 1 : 0
+  count   = length(var.aws_service_control_policies.aws_deny_root_user_ous) > 0 ? 1 : 0
   name    = "LandingZone-DenyRootUser"
   content = file("${path.module}/files/organizations/deny_root_user.json")
   tags    = var.tags
@@ -64,7 +68,7 @@ resource "aws_organizations_policy" "deny_root_user" {
 
 resource "aws_organizations_policy_attachment" "deny_root_user" {
   for_each = {
-    for ou in data.aws_organizations_organizational_units.default.children : ou.name => ou if contains(var.aws_deny_root_user_ous, ou.name)
+    for ou in data.aws_organizations_organizational_units.default.children : ou.name => ou if contains(var.aws_service_control_policies.aws_deny_root_user_ous, ou.name)
   }
 
   policy_id = aws_organizations_policy.deny_root_user.0.id